Added second pass at refactoring BootstrapEstimator API for improved flexibility, both for the client programmer, as well as the library authors. Fixed nextBinomial() API, supposed to return an int instead of a  double.

+++ src/java/edu/uab/ssg/mixomatic/power/jmsl/RandomAdapter.java
@@ -14,7 +14,7 @@ public final class RandomAdapter implements BootstrapEstimator.RandomNumberGener
 		this.random = new Random();
 	}
 
-	public double nextBinomial(int n, double p) {
+	public int nextBinomial(int n, double p) {
 		return random.nextBinomial(n, p);
 	}
